requirements.txt
# Dependencias principales
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.0
folium>=0.15.0
streamlit-folium>=0.15.0
requests>=2.31.0
plotly>=5.18.0
notion-client>=2.3.0
xlsxwriter>=3.1.0

# Autenticación
streamlit-authenticator>=0.2.1
bcrypt>=4.0.1
PyYAML>=6.0.1

# Dependencias para producción
gunicorn>=20.1.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
